Create new layout component
This is just a shell at the moment but is going to be the basis for a common page layout.
I want to switch to absolute paths rather than relative paths for imports so the config files have been updated to point at the source folder as the base path.