Add content for my education for CV
This is the first step in pulling out the content into it's own files.
Decided to create a new qualification type for this as I wasn't happy using educations.
In the end I made some tweaks anyway as I found that creating tests at this level made me really think about what was being stored and why.
Still a bit torn as to whether tests are rqeuired like this for content but didn't take long and I feel this also covers the type pretty well too.